Revert "moving inner classes and enums to separate files"
This reverts commit 55c596dbb2f602824e3e497bb07a4bf0ea7c5eed.
