--- src/main/java/org/meltzg/jmlm/db/UniqueType.java
@@ -1,7 +1,0 @@
-package org.meltzg.jmlm.db;
-
-public enum UniqueType {
-    NOT,
-    PRIMARY_KEY,
-    UNIQUE
-}
